backend/requirements.txt
fastapi
uvicorn
weaviate-client==4.*
python-dotenv
docling
python-multipart
